refactor: Extract share button
