+++ src/components/share-button.tsx
@@ -1,0 +1,131 @@
+import { t } from "@lingui/macro";
+import { Box, Button, Input, Link, Typography } from "@mui/material";
+import { useRef } from "react";
+
+import { TooltipBox } from "src/components/charts-generic/interaction/tooltip-box";
+import { useDisclosure } from "src/components/use-disclosure";
+import { useOutsideClick } from "src/components/use-outside-click";
+import { Icon } from "src/icons";
+import { copyToClipboard } from "src/lib/copy-to-clipboard";
+
+const ShareButton = () => {
+  const { isOpen, open, close } = useDisclosure();
+  const {
+    isOpen: hasInputFocus,
+    open: setFocusOn,
+    close: setFocusOff,
+  } = useDisclosure();
+  const tooltipBoxRef = useRef<HTMLDivElement>(null);
+  const linkRef = useRef<HTMLAnchorElement>(null);
+  const mouse = useRef({ x: 0, y: 0 });
+
+  const handleClick = () => {
+    open();
+    const linkRect = linkRef.current?.getBoundingClientRect() || {
+      x: 0,
+      y: 0,
+      width: 0,
+    };
+    Object.assign(mouse.current, {
+      x: linkRect.width ?? 0,
+      y: linkRect.y ?? 0,
+    });
+  };
+
+  useOutsideClick(tooltipBoxRef, () => {
+    close();
+    setFocusOff();
+  });
+
+  const { isOpen: hasCopied, setIsOpen: setCopied } = useDisclosure();
+  const handleClickCopyButton = async () => {
+    await copyToClipboard(window.location.toString());
+    setCopied(true);
+    setTimeout(() => setCopied(false), 3000);
+  };
+
+  return (
+    <>
+      <Link
+        variant="body2"
+        color="text.primary"
+        ref={linkRef}
+        onClick={handleClick}
+        sx={{
+          cursor: "pointer",
+          display: "flex",
+          alignItems: "center",
+          gap: 1,
+        }}
+      >
+        <Icon name="share" size={20} />
+        {t({ id: "map.share", message: "Share" })}
+      </Link>
+      {isOpen && (
+        <TooltipBox
+          ref={tooltipBoxRef}
+          placement={{ x: "center", y: "top" }}
+          margins={{ top: 0, bottom: 0, left: 0, right: 0 }}
+          x={mouse.current.x}
+          y={0}
+          interactive
+        >
+          <Box
+            display="flex"
+            mb={2}
+            alignItems="center"
+            justifyContent="space-between"
+          >
+            <Typography variant="h6">URL</Typography>
+            <Typography variant="caption" color="success">
+              {hasCopied
+                ? t({ id: "share.url-copied", message: "URL copied ✅" })
+                : ""}
+            </Typography>
+          </Box>
+          <Box
+            sx={{
+              borderStyle: "solid",
+              boxSizing: "border-box",
+              borderWidth: 1,
+              borderColor: "monochrome.300",
+              outline: hasInputFocus ? "2px solid" : "none",
+              outlineColor: "primary.main",
+              display: "flex",
+              alignItems: "center",
+              borderRadius: 6,
+              overflow: "hidden",
+            }}
+          >
+            <Input
+              onFocus={setFocusOn}
+              onBlur={setFocusOff}
+              sx={{
+                width: 300,
+                border: "none",
+                height: "100%",
+                "&:focus": { border: "none", outline: 0 },
+              }}
+              value={window.location.toString()}
+            />
+            <Button
+              color="secondary"
+              onClick={handleClickCopyButton}
+              sx={{
+                width: "3rem",
+                height: "3rem",
+                display: "flex",
+                alignItems: "center",
+                justifyContent: "center",
+              }}
+            >
+              <Icon name="duplicate" />
+            </Button>
+          </Box>
+        </TooltipBox>
+      )}
+    </>
+  );
+};
+
+export default ShareButton;
